reporting-app/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
python-dotenv
